ansible/roles/django/files/DjangoApi/requirements.txt

distlib==0.3.4
distro==1.7.0
Django==4.1.4
django-allauth==0.51.0
django-contact-form==2.0.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-modeltranslation==0.18.7
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
Pillow==9.2.0
pipenv==2022.10.25
virtualenv==20.15.1
